% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_deviation.R
\name{standard_deviation}
\alias{standard_deviation}
\alias{standard_deviation,numeric-method}
\title{Calculate Standard Deviation (S4)}
\usage{
standard_deviation(x, na.rm = FALSE, for.sample = TRUE)

\S4method{standard_deviation}{numeric}(x, na.rm = FALSE, for.sample = TRUE)
}
\arguments{
\item{x}{A \code{numeric} object.}

\item{na.rm}{Logical, whether to remove NA values.}

\item{for.sample}{Logical, whether to calculate sample standard_deviation (default: TRUE).}
}
\value{
A numeric value: the standard deviation.
}
\description{
Computes the standard deviation of a numeric object.
}
\section{Functions}{
\itemize{
\item \code{standard_deviation(numeric)}: Method for numeric

}}
\examples{
v <- c(1, 2, 3, 4, 5)
standard_deviation(v)
standard_deviation(v, for.sample = FALSE)
}
