% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_r.R
\name{table_r}
\alias{table_r}
\alias{table_r,vector,missing-method}
\alias{table_r,vector,vector-method}
\title{Frequency Table for StatVector}
\usage{
table_r(
  x,
  y = NULL,
  type = c("absoluts", "relatives"),
  var.names = NULL,
  digits = 3,
  sum_of_relative_frequency = 100
)

\S4method{table_r}{vector,missing}(
  x,
  y = NULL,
  type = c("absoluts", "relatives"),
  var.names = NULL,
  digits = 3,
  sum_of_relative_frequency = 100
)

\S4method{table_r}{vector,vector}(
  x,
  y = NULL,
  type = c("absoluts", "relatives"),
  var.names = NULL,
  digits = 3,
  sum_of_relative_frequency = 100
)
}
\arguments{
\item{x}{A vector object representing the main variable.}

\item{y}{An optional second vector object for cross-tabulation.}

\item{type}{A character vector indicating the type of frequencies to compute: "absoluts" or "relatives" (default).}

\item{var.names}{Optional character vector of length 1 (for one variable) or 2 (for two variables), naming the variables.}

\item{digits}{Number of digits to round the output values (default is 3).}

\item{sum_of_relative_frequency}{The base total for percentage calculations (default is 100).}
}
\value{
A matrix with frequency and/or percentage information.
}
\description{
This generic function computes frequency tables for objects of class StatVector.
It can generate frequency tables for one or two variables.
}
\section{Functions}{
\itemize{
\item \code{table_r(x = vector, y = missing)}: Method for single vector object

\item \code{table_r(x = vector, y = vector)}: Method for two vector objects

}}
